Amélioration des commandes

+++ src/application/modeles/Moissonneuse.java
@@ -2,7 +2,7 @@ package application.modeles;
 
 import application.classes.ElementPair;
 import application.classes.Point;
-import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.SimpleFloatProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 
 /**
@@ -12,26 +12,26 @@ public class Moissonneuse extends Vehicule {
 
     private SimpleIntegerProperty capacite_tremis; // Capacité du trémis en Litre
     private SimpleIntegerProperty capacite_reservoir; // Capacité du réservoir en Litre
-    private SimpleDoubleProperty largeur; // Largeur en Mètre
-    private SimpleDoubleProperty hauteur; // Hauteur en Mètre
-    private SimpleDoubleProperty taille_coupe; // Taille de coupe en Mètre
+    private SimpleFloatProperty largeur; // Largeur en Mètre
+    private SimpleFloatProperty hauteur; // Hauteur en Mètre
+    private SimpleFloatProperty taille_coupe; // Taille de coupe en Mètre
     private SimpleIntegerProperty conso_fonctionnement; // Consommation en fonctionnement en Litre par Heure
     private SimpleIntegerProperty conso_route; // Consommation sur route en Litre par Heure
-    private SimpleDoubleProperty poids; // Poids en Tonne
+    private SimpleFloatProperty poids; // Poids en Tonne
 
     private final static String TYPE = "Moissonneuse";
 
-    public Moissonneuse(int id, String marque, String modele, String etat, Point position, int capacite_tremis, int capacite_reservoir, double largeur, double hauteur, double taille_coupe, int conso_fonctionnement, int conso_route, double poids) {
+    public Moissonneuse(int id, String marque, String modele, String etat, Point position, int capacite_tremis, int capacite_reservoir, float largeur, float hauteur, float taille_coupe, int conso_fonctionnement, int conso_route, float poids) {
         super(id, TYPE, marque, modele, etat, position);
 
         this.capacite_tremis = new SimpleIntegerProperty(capacite_tremis);
         this.capacite_reservoir = new SimpleIntegerProperty(capacite_reservoir);
-        this.largeur = new SimpleDoubleProperty(largeur);
-        this.hauteur = new SimpleDoubleProperty(hauteur);
-        this.taille_coupe = new SimpleDoubleProperty(taille_coupe);
+        this.largeur = new SimpleFloatProperty(largeur);
+        this.hauteur = new SimpleFloatProperty(hauteur);
+        this.taille_coupe = new SimpleFloatProperty(taille_coupe);
         this.conso_fonctionnement = new SimpleIntegerProperty(conso_fonctionnement);
         this.conso_route = new SimpleIntegerProperty(conso_route);
-        this.poids = new SimpleDoubleProperty(poids);
+        this.poids = new SimpleFloatProperty(poids);
 
         getInformations().add(new ElementPair("Capacité trémis", capacite_tremis + " Litres"));
         getInformations().add(new ElementPair("Capacité réservoir", capacite_reservoir + " Litres"));
@@ -59,7 +59,7 @@ public class Moissonneuse extends Vehicule {
         this.capacite_reservoir.set(capacite_reservoir);
     }
 
-    public double getLargeur() {
+    public float getLargeur() {
         return largeur.get();
     }
 
@@ -67,7 +67,7 @@ public class Moissonneuse extends Vehicule {
         this.largeur.set(largeur);
     }
 
-    public double getHauteur() {
+    public float getHauteur() {
         return hauteur.get();
     }
 
@@ -75,7 +75,7 @@ public class Moissonneuse extends Vehicule {
         this.hauteur.set(hauteur);
     }
 
-    public double getTaille_coupe() {
+    public float getTaille_coupe() {
         return taille_coupe.get();
     }
 
@@ -99,7 +99,7 @@ public class Moissonneuse extends Vehicule {
         this.conso_route.set(conso_route);
     }
 
-    public double getPoids() {
+    public float getPoids() {
         return poids.get();
     }
